requirements.txt
streamlit
torch
torchvision
numpy
opencv-python-headless
Pillow
albumentations